+++ package.json
@@ -15,10 +15,7 @@
     "@turf/turf": "^6.5.0",
     "axios": "^1.6.2",
     "core-js": "^3.29.0",
-    "howler": "^2.2.4",
-    "moment": "^2.29.4",
-    "ol": "^8.2.0",
-    "pinia": "^2.0.0",
+    "moment": "^2.30.1",
     "roboto-fontface": "*",
     "vue": "^3.5.12",
     "vue-draggable-plus": "^0.5.3",
@@ -28,17 +25,19 @@
   "devDependencies": {
     "@babel/types": "^7.21.4",
     "@types/howler": "^2.2.11",
-    "@types/node": "^18.15.0",
-    "@vitejs/plugin-vue": "^4.0.0",
+    "@types/moment": "^2.11.29",
+    "@types/node": "^24.6.0",
+    "@types/vue-router": "^2.0.0",
+    "@vitejs/plugin-vue": "^6.0.1",
     "@vue/eslint-config-typescript": "^12.0.0",
     "eslint": "^8.37.0",
     "eslint-plugin-vue": "^9.3.0",
     "sass": "^1.60.0",
     "typescript": "^5.3.3",
     "unplugin-fonts": "^1.0.3",
-    "vite": "^5.2.8",
-    "vite-plugin-pwa": "^0.20.5",
+    "vite": "^7.1.7",
+    "vite-plugin-pwa": "^1.0.3",
     "vite-plugin-vuetify": "^2.0.3",
     "vue-tsc": "^2.0.11"
-}
+  }
 }
