--- src/store/index.ts
@@ -1,4 +1,0 @@
-// Utilities
-import { createPinia } from 'pinia'
-
-export default createPinia()
