--- src/ws.vue
@@ -1,15 +1,0 @@
-<template>
-    <h2>Admin Panel</h2>
-</template>
-
-
-<script lang="ts">
-import { useRouter } from 'vue-router';
-
-const router = useRouter() 
-export default {
-  name: "AdminPanel",
-};
-
-
-</script>
